requirements.txt
Flask
bcrypt
psycopg2-binary
Flask-SQLAlchemy
Flask-JWT-Extended
mysqlclient
Flask-Migrate
python-dotenv
gunicorn
